filters: Adjust callback API for flags/errors
It's time to expose the full semantics already in use by the backend
to our filters, by exposing flags and an explicit parameter for
tracking the error value to return to the client.  This is an
incompatible API change, and all existing filters are updated to
match the new semantics.

Of note: access to flags means the cache filter can now behave
correctly in the face of FUA semantics (a FUA write must write
through, even when the cache mode is writeback, rather than
waiting for a subsequent flush).  However, more work to more
fully support FUA through the stack will come in later patches.

I considerered a possible boilerplate reduction option of having
a filter that returns -1 but does not set *err to cause
filters.c to populate *err with the current value of errno; but
for now, manually setting *err at all sites makes it obvious
how much this API change affects.

Signed-off-by: Eric Blake <eblake@redhat.com>