src: Let internal.h handle common includes
There is no need to have all of the .c files under src/ include
nbdkit-plugin.h, since internal.h already handles this.  Also,
this will make it easier when an upcoming patch updates the
public header to respond to a user-defined selection between
API levels; we want all of the files in src/ to see the latest
API, but it is easier to do this from the central location of
internal.h instead of repeated in each .c file.

Signed-off-by: Eric Blake <eblake@redhat.com>