filters: Add nozero filter
Sometimes, it's nice to see what a difference it makes in timing
or in destination file size when sparse handling is enabled or
disabled.  Add a new filter that makes it trivial to disable
write zero support, both at the client side (the feature is not
advertised, so the client must do fallbacks) and at the server
side (the feature is always emulated, rather than using any
fast paths built into the plugin).

The log filter makes it easy to test the difference between the
two modes.

Signed-off-by: Eric Blake <eblake@redhat.com>