todo: Mention possibility of caching .can_FOO callbacks
Recent patches clarified documentation to point out that within
the life of a single connection, the .can_FOO helpers should
return consistent results, and that callers may cache those
results.  But at least in the case of .can_fua, we aren't really
caching things; depending on the overhead involved, calling out
to the plugin's .can_fua on every .pwrite with FUA requested
may be noticeable overhead compared to caching it.  Any cache
must not be a static variable, as it can differ between
connections.

Signed-off-by: Eric Blake <eblake@redhat.com>