tests: Add coverage of zero in language bindings
The existing test of language bindings was not covering any
use of NBD_CMD_WRITE_ZEROES, making it harder to test changes
to the plugin's zero callback.  Recent Linux kernels are now
smart enough to turn fallocate(2) with FALLOC_FL_ZERO_RANGE on
a block device into a SCSI WRITE SAME request, which qemu in
turn converts into an NBD_CMD_WRITE_ZEROES.  Since libguestfs
does not (yet?) directly have an API for this, we have to use
the backdoor of guestfs_debug() and fallocate(1).

On older kernels, with missing fallocate(2) support on block
devices, fallocate(1) fails and no WRITE_ZEROES request is
emitted, but such failure is silently ignored; in fact, even
when the plumbing is NOT missing, if a 'zero' callback causes
an 'EIO' failure, it still won't stop 'make check'.  Thus,
the ONLY way to check that 'zero' is working is to use a
modern system that can plumb everything through, then manually
read the test logs to see what happened when 'zero' was
invoked.  Still, even this manual debugging is a significant
improvement over no coverage at all, and provides the
framework used for testing the next few patches that will be
improving language bindings.

Signed-off-by: Eric Blake <eblake@redhat.com>