tests: Run tests that don't require libguestfs
A bit of refactoring to the HAVE_LIBGUESTFS conditional,
coupled with the addition of a LIBGUESTFS_TESTS intermediate
list, allows us to run a few more tests on CentOS 6 (adding
some tests of command-line behavior and filters that was
previously completely skipped).

For the two tests that we can't run, using check_PROGRAMS
still causes those programs to try to compile (which leads to
compile errors); but using EXTRA_PROGRAMS keeps automake
silent without getting those sources in the way, while still
distributing the sources for someone motivated to fix the
tests.

Compared to the previous patch (14/14 pass and 5 skip), I now
have 15/18 pass and 9 skip.

Signed-off-by: Eric Blake <eblake@redhat.com>