plugins: Fix can_fua default to obey documentation
The documentation for plugins .can_fua added in commit 20db811e
has conflicting information: the first paragraph stated that we
can call .flush even when .can_flush returns false, while the
last claims we check .can_flush.  But the documentation for
.flush itself was consistent at documenting our intent that
it can be called even when .can_flush returns false; so fix
the code to actually do that by getting rid of the spurious
call to .can_flush, and tightening the documentation.

Signed-off-by: Eric Blake <eblake@redhat.com>