connections: Protect open and close callbacks with the request lock.
The thread model SERIALIZE_ALL_REQUESTS claims that although multiple
handles may be open at the same time,

    "[...] data requests
     are serialized so that for the plugin as a whole only one
     read/write/etc request will be in progress at any time."

The text seems to apply to only datapath requests but goes on to say:

    "This is a useful setting if the library you are using is not
     thread-safe.  However performance may not be good."

If the plugin calls into a non-thread-safe library during open() or
close() (as the Python plugin does) then you are not protected.

We could change every plugin, but it seems better to extend the
protection of SERIALIZE_ALL_REQUESTS so it also applies to open() and
close() callbacks, to reduce confusion.

This fixes two problems in the Python plugin if the user-defined
close() callback takes a long time to run (for example it is doing
cleanups).  See:

https://bugzilla.redhat.com/show_bug.cgi?id=1566516
https://bugzilla.redhat.com/show_bug.cgi?id=1566522
(cherry picked from commit 9e6d990f974b05697f9d917ba14f6df94f68cedb)