file: Fix zero/trim with block device
When using block device on RHEL 7.5, file plugin fails to zero with this
error (copied from strace):

[pid 39551] fallocate(8, FALLOC_FL_ZERO_RANGE, 1536, 64000) = -1 ENODEV (No such device)

This is expected error according to the manual:

ENODEV fd does not refer to a regular file or a directory.  (If fd is a
pipe or FIFO, a different error results.)

Treat this error as EOPNOSUPP.

Tested only on Fedora 28; I don't know how to build nbdkit on RHEL, but
the change is pretty trivial.

(cherry picked from commit 8cfafced9acdbf6980b8b6fe108885f8da27ec04)