filters: Print filter name in debugging messages.
nbdkit_debug implicitly prints the plugin name but not the filter
name.  This leads to confusing debug messages such as:

  nbdkit: pattern[1]: debug: prepare
  nbdkit: pattern[1]: debug: prepare

In this instance, two different prepare functions were called in
different filters, but neither filter name was shown.  The plugin name
(pattern) is shown, but that is irrelevant as prepare is a method only
present in filters.

By printing the filter name in the debug message we improve the output:

  nbdkit: pattern[1]: debug: truncate: prepare
  nbdkit: pattern[1]: debug: map: prepare

Another example of the improved output is:

  nbdkit: pattern[1]: debug: truncate: can_flush     # filter
  nbdkit: pattern[1]: debug: map: can_flush          # filter
  nbdkit: pattern[1]: debug: can_flush               # plugin

Also this changes a few filenames in messages to simple filter names.
The filenames are very long and since a filter can only be loaded once
printing the filename more than once is a waste of space.

(cherry picked from commit e7d8da6af9e307364460f02abc53ba18952a77e1)