src: --log option: Implement logging to stderr or syslog.
Use --log=syslog to log errors to the system log, or --log=stderr to
continue the previous behaviour of logging to stderr.

The default is to log to stderr, unless we forked into the background
after which we log to syslog.

This fixes the original broken --log support added in nbdkit 1.5.4
(commit 7c0b4f8d5f84bc6d21666a8b1ff1738e3ed3e172).