connections: Remove confusing debug message.
In discussion on a bug report we were both confused by the debug
message:

  nbdkit: python[1]: debug: connection cleanup with final status 1

What is the meaning of the status here?  Turns out it's essentially an
internal value of no significance.

Also the message as a whole is not really significant, just indicating
that the function handle_single_connection is returning, which we can
deduce anyway.

Delete the whole debug message as it's confusing.