plugins: Implement magic config key.
This extends the concept of the magic script parameter.  In existing
nbdkit plugins if the first parameter is "bare" (does not contain
'='), then it is parsed magically to allow scripting languages to
work:

  nbdkit perl foo.pl

is parsed as:

  nbdkit perl script=foo.pl

This generalises the concept, allowing a plugin.magic_config_key to be
defined.  If this is non-NULL then any bare parameters on the command
line:

  nbdkit file foo

where file_plugin.magic_config_key = "file" is parsed as:

  nbdkit file file=foo

If magic_config_key == NULL then the previous behaviour is used.

There is a small (but hopefully not noticable) change to the behaviour
of ‘--dump-plugin’.  It is now handled after all command line
parameters have been parsed but before the .config_complete method is
called (whereas previously it would be called after the first command
line parameter was parsed).

Note this change only applies to plugin parameters.