tests: Pick a different starting port for test-tls and test-tls-psk.
These two tests start simultaneously and use a not great method to
pick a random unused port.  However they both start searching from
port 50000, and tend to race to pick the same port, thus conflicting.

We should try to find a better method of finding an unused port, but
in the absence of an obviously better method, use a different starting
port for the search.