sh: Add load, unload methods, $tmpdir.
Previously we didn't really have a good plan for how scripts were
supposed to store state between invocations.  This changes the sh
plugin so that it creates a temporary directory at the start and
deletes it when nbdkit exit.  This path is passed to the shell scripts
in $tmpdir, so they can use it for (temporary) state.

This also adds load and unload method calls into the script.

This also improves documentation about how to use a temporary
directory for per-handle state.  (You could do this before, but using
$tmpdir means handles can be cleaned up by nbdkit on exit too).