core: Unload plugins and filters in --help, --version and --dump-plugin.
Make sure we call backend->free (which calls the unload methods of the
plugins and filters) on successful exit paths.  It was not being
called for --help, --version or --dump-plugin, which meant that any
external effects of the load method could not be undone by unload.

One particular case where this was observable was with
nbdkit-sh-plugin, where doing:

  $ nbdkit sh --dump-plugin

previously left an empty /tmp/nbdkitsh* directory ($tmpdir) around.

We still don't attempt to unload the plugins along error paths.  For
example doing this:

  $ nbdkit sh
  nbdkit: error: missing script parameter

still leaves the empty directory around.  There are many such error
paths and it may not even be safe to call the unload method after the
plugin has returned an error (eg. the internal state of the plugin may
not be well defined).