tests: Move common code for testing every plugin to tests/functions.sh.
This resurrects the unused tests/functions.sh file (although now we
need to generate it from tests/functions.sh.in).  Put the common code
for running a test against every plugin here.

Because of the previous commits we can now use the plugins list
directly from configure.ac instead of needing to use weird shell
script, although we still need to preserve the test that the plugin
was built so that the tests continue to work if a plugin is disabled
or not built through missing dependencies.