tests: Add a helper function which waits for nbdkit to start up.
This assumes bashisms, but bash is required to run the tests.

This is mostly refactoring.  However the changes (simplifications) are
quite substantial:

 - Since the new start_nbdkit helper function cleans up nbdkit on
   exit, most scripts no longer need to deal with the pid or kill the
   pid in the cleanup function.

 - As a result, cleanup functions are radically simpler, and often
   disappear completely.

 - Also removed the comment "# The cleanup() function is called
   implicitly on exit" passim partly because the cleanup functions
   have mostly been removed and partly because it should be clear from
   looking at functions.sh.

There is one additional change: In the test-memory*.sh tests, nbdkit
used to run in the foreground, but that seems to be a consequence of
some left over debugging.  I changed it so they work like the other
tests.