build: add a simple test for iconv
It's used by one plugin.

There is a proper iconv.m4 available but it makes autoconf fail
if we don't bring in more of gettext, which wants to add 50 files around.

Original Author: François Revol <revol@free.fr>

RWMJ: I modified this to make the iconv dependency optional.
If unavailable, the floppy plugin is disabled.