main: Use new bool parser for --tls
Our use of --tls is a superset of boolean because we add a
force mode; but once we have checked for that, it's more
consistent if we accept the same set of other boolean values
as anything else (in this case, adding 'yes/no' and 'true/false',
as well as consistently allowing case-insensitivity).  The error
message for an unrecognized spelling changes from:

$ nbdkit --tls huh
nbdkit: --tls flag must be off|on|require

to:

$ nbdkit --tls huh
nbdkit: error: could not decipher boolean (huh)

but I don't think that hurts too much.

Signed-off-by: Eric Blake <eblake@redhat.com>