main: Saner newline printing during --help
We are inconsistent on whether our plugins and filters end their
description and config_help in a newline; as a result, things like

  nbdkit --help --filter=log --filter=nozero --filter=delay memory

have inconsistent spacing between sections.  Rather than documenting
whether the user must have/avoid a trailing newline, and auditing
our existing plugins/filters to comply, it's nicer to just supply
one ourselves automatically and only when it is missing.

Signed-off-by: Eric Blake <eblake@redhat.com>