tests: ocaml: Simplify this test so it runs faster.
This test took 54 seconds to run on my laptop which is quite long.
However without KVM it takes over 15 minutes to run (I don't actually
know how long because I got bored and killed it).

The long running of the test (with KVM) is normally hidden by running
all the tests in parallel.  However when building in Koji we don't use
parallelism, nor KVM, and as a result this test “blocks” the test
suite for a very long time.

We don't have any particular evidence that this test is really testing
anything useful about the OCaml bindings, so I have simplified it and
it now runs in about 58 seconds without KVM and only a few seconds
with KVM.