sh: Fix failing test.
Very occasionally the shell plugin test can fail with:

nbdkit: sh[1]: debug: get_size
nbdkit: sh[1]: error: could not parse size string ()
nbdkit: sh[1]: error: /home/rjones/d/nbdkit/tests/test-shell.sh: could not parse output from get_size method:

(This is most easily observed on a highly loaded machine)

I believe this is down to the script writing data (the size string)
and closing the socket quickly, and on the nbdkit side the poll(2)
system call receiving revents = POLLIN & POLLHUP.  Because we tested
for POLLHUP first and closed the socket, we didn't read the data on
the socket, thus effectively seeing an empty string.

Instead of this check POLLHUP second.  Also we must check if read(2)
returns 0 in the POLLIN case to indicate end of input.

For consistency the same change is made for the stderr socket.