partitioning: Rename crc32 to efi_crc32 to avoid conflict with zlib.
zlib exports a symbol called crc32, which is also the name of our
EFI-specific CRC calculation function.  As a result if we link nbdkit
or the plugin with zlib (even indirectly) then we may end up using the
zlib symbol, which causes a segmentation fault.  Renaming our function
avoids this.  I also renamed the source files to make the name less
generic, since this function really only works for EFI CRCs, not for
general CRCs.

This only affects RHEL 7.  I believe that's because Fedora uses a flag
such as --as-needed or similar.