configure: Fix formatting of --help output.
Because of m4/autoconf/bash voodoo we must always put [...] as
“tightly” as possible around AS_HELP_STRING, otherwise the indentation
of the ‘./configure --help’ output is wrong.

I also tried to make the capitalization consistent, and improved the
indentation of the configure.ac program itself.

Before this patch we had weird indentation problems, as shown in the
extract below.

┈┈┈┈✂┈┈┈┈
  --without-iconv         Don't try to link against iconv

          --with-tls-priority     default TLS session priority string [default=NORMAL]

      --without-curl          disable curl plugin [default=check]

      --without-iso           disable iso plugin [default=check]
┈┈┈┈✂┈┈┈┈

After this patch all the output lines up, capitalization is
consistent, etc:

┈┈┈┈✂┈┈┈┈
  --without-iconv         don't try to link against iconv [default=check]
  --with-tls-priority     default TLS session priority string [default=NORMAL]
  --without-curl          disable curl plugin [default=check]
  --without-iso           disable iso plugin [default=check]
┈┈┈┈✂┈┈┈┈