log: Silence harmless Coverity warning
Coverity warned that the value stored to 's' is never read;
but as we just documented that our datestamp printing is
best effort, and we really shouldn't be overflowing the buffer,
there's really nothing further to do with the snprintf()
result.  On the other hand, we can adjust the strftime() call
such that the existing assert now ensures we are not setting
ourselves up for unneeded truncation.

Reported-by: Richard W.M. Jones
Signed-off-by: Eric Blake <eblake@redhat.com>