sh: Make ‘open’ method optional (returning h == "" if omitted).
When writing a very minimal plugin to simulate a disk with bad blocks,
I ended up with:

  case "$1" in
      open) ;;
      get_size) echo 64M ;;
      pread)
          if [ $4 -le 100000 ] && [ $(( $4+$3 )) -gt 100000 ]; then
              echo EIO Bad block >&2
              exit 1
          else
              dd if=/dev/zero count=$3 iflag=count_bytes
          fi ;;
      *) exit 2 ;;
  esac

While this is very small and illustrates the benefits of the ‘sh’
plugin for testing, the open method is obviously redundant.

Allow it to be omitted, with the handle being returned as an empty
string in this case.  (Note this doesn't stop any existing code from
being valid, nor does it prevent anything that was possible before
eg. you can still define the method if you want to return an error in
some cases).

Also document that the ‘get_size’ and ‘pread’ methods are required.

(cherry picked from commit d2571b330b23a36f1161b095df1ae70735555442)