openbsd: Check for ‘seq’ command in configure; replace when using bash.
OpenBSD doesn't have ‘seq’ (but if you install coreutils it has
‘gseq’).  This annoying difference means we can’t reliably use the
command everywhere.

In tests/Makefile we don’t know what shell we’re using, so detect if
‘seq’ or ‘gseq’ is available and use $(SEQ) there.

In test scripts (which are required to use bash) we can use the
bash-specific alternative ‘{M..N}’, and that may be faster too.