connections: Be less noisy when client sends NBD_OPT_ABORT
At least 'nbd-client -l localhost' is a good demonstration of a
client that legitimately sends NBD_OPT_ABORT - it exists only
to enumerate the exports advertised by the server, rather than
intending to connect to any of those exports.  As such, we should
downgrade the message for a client telling us it is going away
to the same status as our debug message for NBD_CMD_DISC, given
that both commands have the same effect at different points in
the protocol.  This stops nbdkit from spamming stderr with

nbdkit: file[1]: error: client sent NBD_OPT_ABORT to abort the connection

messages due to a client that doesn't want to connect after all.

Signed-off-by: Eric Blake <eblake@redhat.com>