main: Avoid fprintf(%m), for BSD builds
printf %m is a glibc extension; using it directly can cause
odd output on BSD. Most of our uses of %m are via nbdkit wrappers,
which will be fixed in the next patch; this was the only %m I
could find in a direct call to the printf family, so it doesn't
hurt to just open-code it.  This particular error is reached
early enough in the process that nbdkit_error() is not a viable
substitute.

Signed-off-by: Eric Blake <eblake@redhat.com>