log: Ensure %m sees correct errno
If a user passes %m in a format string, but we do any
intermediate processing that might corrupt errno, then we
risk changing the message that the user intended to be
printed to instead refer to our changed state of errno.
Hoist the saving of errno to occur before any other risky
calls, and restore errno before printing a user-supplied
format.

Signed-off-by: Eric Blake <eblake@redhat.com>