sh: Don't let child inherit SIGPIPE ignored
While nbdkit itself must run with SIGPIPE ignored, many applications
expect to inherit SIGPIPE in the default state. What's worse, POSIX
states that a non-interactive shell script cannot use 'trap' to
undo an inherited SIG_IGN on SIGPIPE.  I have seen several bug
reports over the years of something that works for a developer but
fails under a CI environment, where the root cause was the CI
leaking an ignored SIGPIPE into the file under test.

Testing this proved to be the more interesting part of the patch.
'yes' is probably the easiest way to generate lots of data and
which behaves differently if SIGPIPE is ignored, but I'm not
certain if coreutils is always installed on BSD machines.  Also
fix a missing redirect to stderr in the probe for $tmpdir.

Signed-off-by: Eric Blake <eblake@redhat.com>