common: Move shared bitmap code to a common library.
The cow and cache filters both use a bitmap mapping virtual disk
blocks to status stored in the bitmap.  The implementation of the
bitmaps is very similar because one was derived from the other when
the filters were implemented.

The main difference is the cow filter uses a simple bitmap (one bit
per block), whereas the cache filter uses two bits per block.

This commit abstracts the bitmap structure into a common library.  The
block size and bits per block are configurable.

This commit is refactoring, except:

- it cures a memory leak (bitmap was not being freed),

- in the cache filter we now mask out existing bits when updating the
  bitmap.