valgrind: Remove --child-silent-after-fork.
Valgrind has two ways to control valgrinding into subprocesses,
--child-silent-after-fork and --trace-children.

--child-silent-after-fork=yes causes tracing to stop when the process
forks.  However in nbdkit we want to continue tracing nbdkit when it
forks itself into the background, so I have removed this option now.

--trace-children=no causes tracing to stop when the program calls one
of the exec(2) functions.  For nbdkit we want this function since we
don't want to trace into subprocesses (eg. when using
nbdkit-sh-plugin).

Note that although we are now tracing into the subprocess, the tests
will still not exit with error if the subprocess leaks memory because
we would need to properly ‘wait -n’ for the subprocess, which
tests/functions.sh does not do.  However a message about the leak will
still get printed in the log.